chore: 完成实验E1和G0的结果整理与配置更新
本次提交包含：
1.  更新两个实验的模型路径为本地离线模型路径
2.  修复KV cache张量哈希计算的精度问题
3.  兼容transformers 5.x新版DynamicCache格式
4.  补充G0实验的显存、精度、量化测试结果报告
5.  新增E1实验的工作负载特征分析报告与轨迹数据
6.  更新判定报告，当前G0已通过4/6项测试

+++ experiments/g0/codec_spike.py
@@ -130,7 +130,7 @@ def collect_blocks(backend, cases: Dict, block_size: int, max_blocks: int = 500)
                     break
                 # 用第一层 key 张量的内容哈希作为指纹
                 k_tensor = block["layer_k"][0]
-                content_hash = hash(k_tensor.cpu().numpy().tobytes())
+                content_hash = hash(k_tensor.cpu().float().numpy().tobytes())
                 if content_hash not in seen_fingerprints:
                     seen_fingerprints.add(content_hash)
                     # 将 block 的 KV 张量移到 CPU，避免 GPU 显存累积
